+++ CHANGELOG.md
@@ -13,6 +13,10 @@ authorization request.
 and the storagerelay scheme.
 
 ### Changed
+- Renamed `http://i.authnomicon.org/oauth2/AuthorizationCodeService` interface
+to `module:@authnomicon/oauth2.AuthorizationCodeService`.
+- Renamed `http://i.authnomicon.org/oauth2/AccessTokenService` interface to
+`module:@authnomicon/oauth2.AccessTokenService`.
 - Renamed `http://i.authnomicon.org/oauth2/authorization/http/ResponseType`
 interface to `module:oauth2orize.RequestProcessor`.
 - Renamed `http://i.authnomicon.org/oauth2/authorization/http/RequestParameters`